# Cache reflection portion of `SettingSource` lookups