Fixes the mod-related overhead portion of https://github.com/ppy/osu/issues/14638. There's still a significant performance issue that will be addressed separately. Still not sure if this is 100% correct direction but seems to work well enough.

Before:

![](https://user-images.githubusercontent.com/191335/132330537-7414e9b2-e243-4836-bbde-211366b0a053.png)

After:

(gone)