@peppy 2021-09-08
To expand on why I'm not sure about this direction: it doesn't change the fact that in the `GetAllMods` calls we are constructing mods for throwaway purposes. This is constructing bindables (and in the case of difficulty adjust it's also binding them together) which is not free.

If we go with this cache, it's still something to keep in mind.