@bdach 2021-09-08
This can't work as written, sorry for misleading (as evidenced by test failures). The binding to the generic method happens at compile time, so all readers that call this on `Mod`s (rather than concrete derived types) won't see any properties. Probably have to fall back to a more conventional `Type`-keyed dictionary setup.