@peppy 2021-09-08
This............. doesn't work. `GetType()` carries the same overhead. Send help.

![20210909 002504 (JetBrains Rider)](https://user-images.githubusercontent.com/191335/132538410-9804d3a4-10f0-429d-8727-a1d18f7ff009.png)