@bdach 2021-09-08
Umm, right... I should have seen that coming but somehow didn't...

The next best thing to use as a cache key (that doesn't touch reflection) would be a `(rulesetId, acronym)` pair, I guess? Which is awkward since it's not currently easily possible to get the first cheaply from a `Mod` instance...

Maybe `RulesetStore` could initialise this setting cache once? Not sure, this is getting into real dicey territory. I *really* don't want to see the reflection get replaced for *some* sort of handrolling but it seems to be increasingly inevitable...