@frenzibyte 2021-09-08
The problem with using such way of caching is that `SettingsSourceAttribute` would then be mods-only code, while I believe it is planned to be used for skin editor and other places (if not used already).

Though looking at it, the method `ResolveTypeHandleInternal` seems to be one that could come from `GetCustomAttribute` instead, as that's the stack trace to it (taking `Drawable.IsLongRunning` as an example source):
![CleanShot 2021-09-08 at 22 01 08](https://user-images.githubusercontent.com/22781491/132569234-2e33c282-0f74-4fa8-a9ea-5ee2c46e56ca.png)

It's also noticable there as `getSettingsSourceProperties` is what shows up in the trace, and the two other methods both call `GetCustomAttribute` as well.

Still not sure why the overhead of `GetCustomAttribute` would remain even when caching the properties though, as `Object.GetType()` is `extern` and calls an internal method implemented within the CLR, so it shouldn't actually lead to `ResolveTypeHandleInternal` as far as I'm seeing here.