@bdach 2021-09-08
upon checking again, `GetType()` should indeed be free.

> Though looking at it, the method `ResolveTypeHandleInternal` seems to be one that could come from `GetCustomAttribute` instead

still doesn't change the fact that the cache initialisation (so first call to `.ToAPIMod()` for a given mod type) will still incur this overhead. so if the goal is to remove it always then the cache may, again, need to be eagerly populated.