@peppy 2021-09-08
it’s not the initialization. my profiling was of course done after ensuring the cache was initialized.