@smoogipoo 2021-09-09
As I said in Discord, it is very likely that we're being misled here and it's not actually reflection that's taking a long time but the fact that we're allocating a lot during this.

```
[Benchmark]
public void GetSettingSources() => mod.GetOrderedSettingsSourceProperties().Consume(new Consumer());

Master:
|            Method |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------ |---------:|---------:|---------:|-------:|----------:|
| GetSettingSources | 27.91 us | 0.076 us | 0.067 us | 0.0916 |      9 KB |

This PR:
|            Method |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------ |---------:|---------:|---------:|-------:|----------:|
| GetSettingSources | 27.99 us | 0.095 us | 0.088 us | 0.0916 |      9 KB |

```