@peppy 2021-09-09
Attempting to drill down into allocations, while there is some overhead on our side, the majority is still coming from `GetCustomAttribute` internals.

![20210909 112747 (Parallels Desktop)](https://user-images.githubusercontent.com/191335/132612330-7f9432cc-5cb9-4230-8e45-b1ab67c46ba7.png)

![20210909 112852 (Parallels Desktop)](https://user-images.githubusercontent.com/191335/132612369-55fa0609-bd89-4520-8ad5-0f8d45afa689.png)

I'll dig deeper and try to isolate a repro case. As you mentioned, it may be async/threading specific, or potentially something to do with the size of the json being parsed here.