@peppy 2021-09-09
Closing this as neither solution is valid. Am investigating today, if I don't come up with a solution I'll open an issue tracking it.