@peppy 2021-09-09
This is actually valid with the embarrassing fix that none of us caught:

base

|                         Method |         Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------- |-------------:|----------:|----------:|-------:|----------:|
|       BenchmarkToModDoubleTime |  27,098.0 ns | 183.44 ns | 171.59 ns | 1.0986 |  12,858 B |
| BenchmarkToModDifficultyAdjust | 120,852.7 ns | 946.96 ns | 839.46 ns | 3.1738 |  37,118 B |
|            BenchmarkGetAllMods |     951.8 ns |   5.85 ns |   5.19 ns | 0.0381 |     440 B |

this PR:
|                         Method |        Mean |     Error |   StdDev |  Gen 0 | Allocated |
|------------------------------- |------------:|----------:|---------:|-------:|----------:|
|       BenchmarkToModDoubleTime |  6,082.8 ns |  58.38 ns | 54.61 ns | 0.3510 |   4,089 B |
| BenchmarkToModDifficultyAdjust | 22,210.1 ns | 117.00 ns | 97.70 ns | 0.7935 |   9,425 B |
|            BenchmarkGetAllMods |    968.0 ns |   4.97 ns |  4.41 ns | 0.0381 |     440 B |