@peppy 2021-09-09
Updated benchmarks (with `GetAllMods` actually being consumed):

master:

|                         Method |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------------------- |----------:|---------:|---------:|-------:|----------:|
|       BenchmarkToModDoubleTime |  26.37 us | 0.096 us | 0.090 us | 1.0681 |     12 KB |
| BenchmarkToModDifficultyAdjust | 118.66 us | 0.885 us | 0.784 us | 3.1738 |     36 KB |
|            BenchmarkGetAllMods |  65.10 us | 0.479 us | 0.448 us | 2.5635 |     29 KB |


this pr:

|                         Method |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------- |----------:|----------:|----------:|-------:|----------:|
|       BenchmarkToModDoubleTime |  6.176 us | 0.0307 us | 0.0272 us | 0.3510 |      4 KB |
| BenchmarkToModDifficultyAdjust | 21.568 us | 0.1429 us | 0.1267 us | 0.7935 |      9 KB |
|            BenchmarkGetAllMods | 15.669 us | 0.1784 us | 0.1669 us | 1.3428 |     15 KB |